+++ apps/backoffice-v2/src/lib/blocks/hooks/useDocumentBlocks/utils/check-can-revision/check-can-revision.ts
@@ -1,7 +1,6 @@
 import { useCaseState } from '@/pages/Entity/components/Case/hooks/useCaseState/useCaseState';
 import { TWorkflowById } from '@/domains/workflows/fetchers';
-import { CommonWorkflowEvent, DefaultContextSchema } from '@ballerine/common';
-import { getPostDecisionEventName } from '@/lib/blocks/components/CallToActionLegacy/hooks/useCallToActionLegacyLogic/useCallToActionLegacyLogic';
+import { CommonWorkflowEvent, CommonWorkflowStates, DefaultContextSchema } from '@ballerine/common';
 import { checkCanMakeDecision } from '@/lib/blocks/hooks/useDocumentBlocks/utils/check-can-make-decision/check-can-make-decision';
 
 export const checkCanRevision = ({
@@ -17,7 +16,8 @@ export const checkCanRevision = ({
   decision: DefaultContextSchema['documents'][number]['decision'];
   isLoadingRevision: boolean;
 }) => {
-  const hasTaskReviewedEvent = !!getPostDecisionEventName(workflow);
+  const isStateManualReview = workflow.state === CommonWorkflowStates.MANUAL_REVIEW;
+
   const hasRevisionEvent = workflow?.nextEvents?.includes(CommonWorkflowEvent.REVISION);
   const canMakeDecision = checkCanMakeDecision({
     caseState,
@@ -25,5 +25,5 @@ export const checkCanRevision = ({
     decision,
   });
 
-  return !isLoadingRevision && canMakeDecision && (hasTaskReviewedEvent || hasRevisionEvent);
+  return !isLoadingRevision && canMakeDecision && (isStateManualReview || hasRevisionEvent);
 };