+++ services/workflows-service/src/workflow/workflow.controller.external.ts
@@ -1,6 +1,6 @@
 import { UserData } from '@/user/user-data.decorator';
 import { UserInfo } from '@/user/user-info';
-import { isRecordNotFoundError } from '@/prisma/prisma.util';
+import { defaultPrismaTransactionOptions, isRecordNotFoundError } from '@/prisma/prisma.util';
 import * as common from '@nestjs/common';
 import { NotFoundException, Query, Res } from '@nestjs/common';
 import * as swagger from '@nestjs/swagger';
@@ -33,6 +33,8 @@ import { Public } from '@/common/decorators/public.decorator';
 import { WorkflowDefinitionService } from '@/workflow-defintion/workflow-definition.service';
 import { CreateCollectionFlowUrlDto } from '@/workflow/dtos/create-collection-flow-url';
 import { env } from '@/env';
+import { PrismaService } from '@/prisma/prisma.service';
+import { BUILT_IN_EVENT, ARRAY_MERGE_OPTION } from '@ballerine/workflow-core';
 
 @swagger.ApiBearerAuth()
 @swagger.ApiTags('external/workflows')
@@ -45,6 +47,7 @@ export class WorkflowControllerExternal {
     protected readonly rolesBuilder: nestAccessControl.RolesBuilder,
     private readonly workflowTokenService: WorkflowTokenService,
     private readonly workflowDefinitionService: WorkflowDefinitionService,
+    private readonly prismaService: PrismaService,
   ) {}
 
   // GET /workflows
@@ -129,29 +132,6 @@ export class WorkflowControllerExternal {
       }
       throw error;
     }
-  }
-
-  // POST /intent
-  @common.Post('/intent')
-  @swagger.ApiOkResponse()
-  @common.HttpCode(200)
-  @swagger.ApiForbiddenResponse({ type: errors.ForbiddenException })
-  @UseCustomerAuthGuard()
-  async intent(
-    @common.Body() { intentName, entityId }: IntentDto,
-    @ProjectIds() projectIds: TProjectIds,
-    @CurrentProject() currentProjectId: TProjectId,
-  ) {
-    const entityType = intentName === 'kycSignup' ? 'endUser' : 'business';
-
-    // @TODO: Rename to intent or getRunnableWorkflowDataByIntent?
-    return await this.service.resolveIntent(
-      intentName,
-      entityId,
-      entityType,
-      projectIds,
-      currentProjectId,
-    );
   }
 
   @common.Post('/run')
@@ -319,24 +299,45 @@ export class WorkflowControllerExternal {
     @common.Body() hookResponse: any,
   ): Promise<void> {
     try {
-      const workflowRuntime = await this.service.getWorkflowRuntimeDataByIdUnscoped(params.id);
-      await this.normalizeService.handleHookResponse({
-        workflowRuntime: workflowRuntime,
-        data: hookResponse,
-        resultDestinationPath: query.resultDestination || 'hookResponse',
-        processName: query.processName,
-        projectIds: [workflowRuntime.projectId],
-        currentProjectId: workflowRuntime.projectId,
-      });
-
-      await this.service.event(
-        {
+      await this.prismaService.$transaction(async transaction => {
+        const workflowRuntime = await this.service.getWorkflowRuntimeDataByIdAndLockUnscoped({
           id: params.id,
-          name: params.event,
-        },
-        [workflowRuntime.projectId],
-        workflowRuntime.projectId,
-      );
+          transaction,
+        });
+
+        const context = await this.normalizeService.handleHookResponse({
+          workflowRuntime: workflowRuntime,
+          data: hookResponse,
+          resultDestinationPath: query.resultDestination || 'hookResponse',
+          processName: query.processName,
+          projectIds: [workflowRuntime.projectId],
+          currentProjectId: workflowRuntime.projectId,
+        });
+
+        await this.service.event(
+          {
+            id: params.id,
+            name: BUILT_IN_EVENT.DEEP_MERGE_CONTEXT,
+            payload: {
+              newContext: context,
+              arrayMergeOption: ARRAY_MERGE_OPTION.REPLACE,
+            },
+          },
+          [workflowRuntime.projectId],
+          workflowRuntime.projectId,
+          transaction,
+        );
+
+        await this.service.event(
+          {
+            id: params.id,
+            name: params.event,
+          },
+          [workflowRuntime.projectId],
+          workflowRuntime.projectId,
+          transaction,
+        );
+      }, defaultPrismaTransactionOptions);
     } catch (error) {
       if (isRecordNotFoundError(error)) {
         throw new errors.NotFoundException(`No resource was found for ${JSON.stringify(params)}`);