+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.5.45
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.45
+  - @ballerine/workflow-core@0.5.45
+
+## 0.5.44
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.44
+  - @ballerine/common@0.7.44
+
 ## 0.5.43
 
 ### Patch Changes